+++ docs/changlog/changelog.md
@@ -2,3 +2,4 @@
 
 ### Added
 - Lazy load list and iterator (Not 100% complete but okay for internal use)
+- Lazy load iterator which only stores the current page of objects